menu management and akses menu done

+++ application/controllers/Menu.php
@@ -17,6 +17,11 @@ class Menu extends CI_Controller {
         $this->form_validation->set_rules('menu', 'Menu', 'required|trim', [
             'required' => 'Menu Harus diisi!!!'
         ]);
+        $this->form_validation->set_rules('urutan', 'Urutan', 'required|trim|numeric|is_unique[admin_menu.urutan]', [
+            'required' => 'Urutan Harus diisi!!!',
+            'numeric' => 'Urutan Harus berupa angka!',
+            'is_unique' => 'Urutan yang diinginkan sudah ada'
+        ]);
 
         if ($this->form_validation->run() === FALSE) {
             $data['title'] = 'Menu Management';
@@ -35,7 +40,10 @@ class Menu extends CI_Controller {
             $this->load->view('templates/modal');
             $this->load->view('templates/footer2');
         } else {
-            $data = ['menu' => $this->input->post('menu', TRUE)];
+            $data = [
+                'menu' => $this->input->post('menu', TRUE),
+                'urutan' => $this->input->post('urutan', TRUE)
+            ];
             $this->Utama_model->insertData('menus', $data);
 
             $this->session->set_flashdata('pesan', '<div class="alert alert-success alert-dismissible fade show" role="alert">Data <strong>Berhasil</strong> ditambah.<button type="button" class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">&times;</span></button></div>');
@@ -46,10 +54,17 @@ class Menu extends CI_Controller {
     public function deleteMenu()
     {
         $data = ['id' => $this->uri->segment(3)];
-        $this->Utama_model->deleteData('menus', $data);
-
-        $this->session->set_flashdata('pesan', '<div class="alert alert-success alert-dismissible fade show" role="alert">Data <strong>Berhasil</strong> dihapus.<button type="button" class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">&times;</span></button></div>');
-        redirect('menu');
+        $cek = $this->Utama_model->getDatas('access_menus_raw', ['menu_id' => $this->uri->segment(3)]);
+        // cek apakah ada data menu di table access menu
+        if (!array_key_exists("status", $cek)) {
+            $this->session->set_flashdata('pesan', '<div class="alert alert-danger alert-dismissible fade show" role="alert">Data <strong>Gagal</strong> dihapus<br>Karena masih adanya data di tabel access menu<button type="button" class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">&times;</span></button></div>');
+            redirect('menu');
+        } else {
+            $this->Utama_model->deleteData('menus', $data);
+    
+            $this->session->set_flashdata('pesan', '<div class="alert alert-success alert-dismissible fade show" role="alert">Data <strong>Berhasil</strong> dihapus.<button type="button" class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">&times;</span></button></div>');
+            redirect('menu');
+        }
     }
 
     public function editMenu()
@@ -57,6 +72,11 @@ class Menu extends CI_Controller {
         // Validasi Menu
         $this->form_validation->set_rules('menu', 'Menu', 'required|trim', [
             'required' => 'Menu Harus diisi!!!'
+        ]);
+        $this->form_validation->set_rules('urutan', 'Urutan', 'required|trim|numeric|is_unique[admin_menu.urutan]', [
+            'required' => 'Urutan Harus diisi!!!',
+            'numeric' => 'Urutan Harus berupa angka!',
+            'is_unique' => 'Urutan yang diinginkan sudah ada'
         ]);
 
         if ($this->form_validation->run() === FALSE) {
@@ -79,9 +99,10 @@ class Menu extends CI_Controller {
         } else {
             $data = [
                 'menu' => $this->input->post('menu', TRUE),
+                'urutan' => $this->input->post('urutan', TRUE),
                 'id' => $this->uri->segment(3)
             ];
-            
+
             $this->Utama_model->updateData('menus', $data);
 
             $this->session->set_flashdata('pesan', '<div class="alert alert-success alert-dismissible fade show" role="alert">Data <strong>Berhasil</strong> diubah.<button type="button" class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">&times;</span></button></div>');
@@ -121,23 +142,92 @@ class Menu extends CI_Controller {
 
             $access_menu_cek = $this->Utama_model->getDatas('access_menus_raw');
 
+            //validasi apabila data sudah ada
             foreach($access_menu_cek as $amc) {
                 if ($level_id === $amc['level_id'] && $menu_id === $amc['menu_id']) {
                     $this->session->set_flashdata('pesan', '<div class="alert alert-danger alert-message" role="alert">Akses Menu Sudah Ada!</div>');
                     redirect('menu/access_menu');
-                } else {
-                    $data = [
-                        'level_id' => $level_id,
-                        'menu_id' => $menu_id
-                    ];
+                }
+            }
+            $data = [
+                'level_id' => $level_id,
+                'menu_id' => $menu_id
+            ];
 
-                    $this->Utama_model->insertData('access_menus_raw', $data);
+            $this->Utama_model->insertData('access_menus_raw', $data);
 
-                    $this->session->set_flashdata('pesan', '<div class="alert alert-success alert-dismissible fade show" role="alert">Data <strong>Berhasil</strong> ditambah.<button type="button" class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">&times;</span></button></div>');
+            $this->session->set_flashdata('pesan', '<div class="alert alert-success alert-dismissible fade show" role="alert">Data <strong>Berhasil</strong> ditambah.<button type="button" class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">&times;</span></button></div>');
+            redirect('menu/access_menu');
+        }
+    }
+
+    public function editAccessMenu()
+    {
+        // Validasi Menu
+        $this->form_validation->set_rules('menu_id', 'Menu', 'required|trim', [
+            'required' => 'Menu Harus diisi!!!'
+        ]);
+        $this->form_validation->set_rules('level_id', 'Level', 'required|trim', [
+            'required' => 'Level Harus diisi!!!'
+        ]);
+
+        if ($this->form_validation->run() === FALSE) {
+            $id = $this->uri->segment(3);
+            $data['title'] = 'Akses Menu Admin';
+    
+            $data['admin'] = $this->Utama_model->getDatas('admins', ['email' => $this->session->userdata('email')])[0];
+
+            $data['access_menu'] = $this->Utama_model->getDatas('access_menus_raw', ['id' => $id])[0];
+
+            $data['menus_by_access_menu'] = $this->Utama_model->getDatas('menus_by_access_menu', ['level_id' => $this->session->userdata('level')]);
+    
+            $this->load->view('templates/header', $data);
+            $this->load->view('templates/sidebar', $data);
+            $this->load->view('templates/topbar', $data);
+            $this->load->view('menu/edit_access_menu', $data);
+            $this->load->view('templates/footer');
+            $this->load->view('templates/modal', $data['title']);
+            $this->load->view('templates/footer2');
+        } else {
+            $level_id = $this->input->post('level_id', true);
+            $menu_id = $this->input->post('menu_id', true);
+
+            $access_menu_cek = $this->Utama_model->getDatas('access_menus_raw');
+
+            foreach($access_menu_cek as $amc) {
+                if ($level_id === $amc['level_id'] && $menu_id === $amc['menu_id']) {
+                    $this->session->set_flashdata('pesan', '<div class="alert alert-danger alert-message" role="alert">Akses Menu Sudah Ada!</div>');
                     redirect('menu/access_menu');
                 }
             }
+
+            $data = [
+                'level_id' => $level_id,
+                'menu_id' => $menu_id,
+                'id' => $this->uri->segment(3)
+            ];
+
+            $this->Utama_model->updateData('access_menus_raw', $data);
+
+            $this->session->set_flashdata('pesan', '<div class="alert alert-success alert-dismissible fade show" role="alert">Data <strong>Berhasil</strong> diubah.<button type="button" class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">&times;</span></button></div>');
+            redirect('menu/access_menu');
         }
+    }
+
+    public function deleteAccessMenu()
+    {
+        $data = ['id' => $this->uri->segment(3)];
+        // $cek = $this->Utama_model->getDatas('access_menus_raw', ['menu_id' => $this->uri->segment(3)]);
+        // cek apakah ada data menu di table access menu
+        // if (!array_key_exists("status", $cek)) {
+        //     $this->session->set_flashdata('pesan', '<div class="alert alert-danger alert-dismissible fade show" role="alert">Data <strong>Gagal</strong> dihapus<br>Karena masih adanya data di tabel access menu<button type="button" class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">&times;</span></button></div>');
+        //     redirect('menu');
+        // } else {
+            $this->Utama_model->deleteData('access_menus_raw', $data);
+    
+            $this->session->set_flashdata('pesan', '<div class="alert alert-success alert-dismissible fade show" role="alert">Data <strong>Berhasil</strong> dihapus.<br>Silahkan tambahkan submenu di menu submenu<button type="button" class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">&times;</span></button></div>');
+            redirect('menu/access_menu');
+        // }
     }
 
 }
